refactor: the api into multiple files for better maintenblity
