+++ dashboard/backend/sql/users.ts
@@ -3,7 +3,7 @@ import { int, sqliteTable, text, primaryKey, index } from "drizzle-orm/sqlite-co
 import { queryCondition, toUndef } from "./sql-helper.js";
 import { eq, and, inArray } from "drizzle-orm/sql/expressions";
 import { AuthProvider, QueryUser, User, UserByProvider, UserStatus } from "@fireproof/core-protocols-dashboard";
-import { DashSqlite } from "./create-handler.js";
+import { DashSqlite } from "../create-handler.js";
 
 export const sqlUsers = sqliteTable("Users", {
   userId: text().primaryKey(),
